Move LegalNameValidator into net.corda.core.internal
